# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Tunnex API
  version: 0.1.0
  description: |
    Control-plane API for Tunnex.io. This spec is the single source of truth:
    the Go server types/handlers and the TypeScript client are generated from it
    (`make generate`), and drift fails the build (`make generate-check`).
servers:
  - url: /
    description: Same-origin (served behind the edge nginx)

# Endpoints require a session cookie OR a CLI bearer credential by default
# (bearer ≡ cookie, S5.1); public ones set `security: []`; cookie-only ops
# (credential minting) override with `security: [{cookieAuth: []}]` + argument.
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/v1/agent/runtime/poll:
    get:
      operationId: pollAgentRuntime
      summary: Poll managed agent configuration
      description: >-
        Machine-only managed-runtime channel. Authenticates with the exact agent runtime bearer and returns
        a configuration only when the desired revision is newer than `applied_revision`; an unchanged
        revision returns 204 with no body. The optional wait is a bounded long poll. Explicit organization
        opt-out is a terminal uniform 401 so the runtime offboards; edition unavailability remains a
        non-terminal 403. No bearer, bootstrap token, private key, token hash, or raw error is ever returned.
      security:
        - runtimeBearer: []
      tags: [agents]
      parameters:
        - name: applied_revision
          in: query
          required: true
          schema: { type: integer, format: int64, minimum: 0 }
        - name: client_version
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 128 }
        - name: wait_seconds
          in: query
          required: false
          description: Bounded long-poll wait requested by the runtime; the server may return sooner.
          schema: { type: integer, minimum: 0, maximum: 60, default: 0 }
      responses:
        "200":
          description: Strictly newer managed configuration.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ManagedAgentConfig" }
        "204":
          description: Desired revision is unchanged; no configuration body is returned.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
        "401": { $ref: "#/components/responses/RuntimeUnauthorized" }
        "403": { $ref: "#/components/responses/Error" }
        default: { $ref: "#/components/responses/Error" }
...
  /api/v1/organizations:
    get:
      operationId: listOrganizations
      summary: List organizations
      description: Lists the organizations the authenticated user belongs to.
      tags: [organizations]
      responses:
        "200":
          description: Organizations.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createOrganization
      summary: Create an organization
      description: |
        Creates an organization. In the open edition this fails with
        `org_limit_reached` once one organization exists; the enterprise build
        allows multiple.
      tags: [organizations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          description: The created organization.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/organizations/{orgId}:
    parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getOrganization
      summary: Get an organization
      tags: [organizations]
      responses:
        "200":
          description: The organization.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        default:
          $ref: "#/components/responses/Error"
    patch:
      operationId: updateOrganization
      summary: Update organization settings (name only; slug is immutable)
      tags: [organizations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationRequest"
      responses:
        "200":
          description: The updated organization.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteOrganization
      summary: Soft-delete an organization (refused while it still owns resources)
      description: >-
        Owner-only (`org:delete`). Refuses with 409 `org_not_empty` while the organization still has
        gateways, sites, devices, Kubernetes clusters or machine credentials — a soft delete leaves them
        running with no organization to manage them from. There is no force flag.
      tags: [organizations]
      responses:
        "204":
          description: Deleted.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: tunnex_session
    # CLI credential (tnx_...). bearer ≡ cookie on every authenticated endpoint;
    # the ONLY exceptions are argued on the ops themselves (cliAuthorize +
    # cliDeviceApprove — credential minting must not be reachable from a bearer).
    bearerAuth:
      type: http
      scheme: bearer
    runtimeBearer:
      type: http
      scheme: bearer
      description: Machine-only hash-backed managed-agent runtime credential; never a user/session bearer.
  headers:
    RequestId:
      description: Correlation ID for this request; also appears in server logs.
      schema:
        type: string
      required: true
  schemas:
...
    Organization:
      type: object
      required: [id, name, slug, pool_cidr, max_agent_identities, managed_agent_runtime_enabled, created_at, updated_at]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        pool_cidr:
          type: string
          description: The org's flat WireGuard address pool (IPv4 CIDR).
        ovpn_enabled:
          type: boolean
          description: "S9.1 D-S9.5-OPTIN: whether this org has opted into OpenVPN. Default false. When false the OpenVPN device type is not offered and the export endpoint refuses (opt_in_required)."
        max_agent_identities:
          type: integer
          format: int32
          minimum: 0
          nullable: true
          description: "F02: organization-wide managed-agent identity quota. null means unlimited; no remaining count is exposed."
        managed_agent_runtime_enabled:
          type: boolean
          description: "F04: explicit organization opt-in for managed runtime synchronization. Default false; a paid licence does not enable it implicitly."
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CreateOrganizationRequest:
      type: object
      required: [name, slug]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          type: string
          minLength: 1
          maxLength: 63
          pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"
    UpdateOrganizationRequest:
      type: object
      required: [name]
      additionalProperties: false
      description: Only the name is mutable; the slug is fixed at creation.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
...
    ManagedAgentConfig:
      type: object
      description: >-
        Secret-free server-owned configuration for one managed agent. Runtime credentials, bootstrap tokens,
        token hashes, and private keys are never configuration fields.
      required: [revision, device_id, org_id, address, gateway_endpoint, gateway_public_key, allowed_ips, dns, persistent_keepalive]
      additionalProperties: false
      properties:
        revision: { type: integer, format: int64, minimum: 1 }
        device_id: { type: string, format: uuid }
        org_id: { type: string, format: uuid }
        address: { type: string, description: The managed agent tunnel address/prefix. }
        gateway_endpoint: { type: string, description: The gateway endpoint in host:port form. }
        gateway_public_key: { type: string, description: The gateway WireGuard public key. }
        allowed_ips:
          type: array
          items: { type: string }
        dns:
          type: array
          items: { type: string }
        persistent_keepalive: { type: integer, minimum: 0, maximum: 65535 }
...
    Error:
      type: object
      description: |
        Standard error envelope for every non-2xx response. Chosen over bare
        RFC 7807 so the correlation `request_id` is a first-class field the SPA
        and CLI can surface directly; `code` is stable for programmatic handling.
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              description: Stable, machine-readable error code.
              example: not_found
            message:
              type: string
              description: Human-readable explanation.
            request_id:
              type: string
              description: Correlation ID; match against server logs.
            details:
              type: array
              description: Field-level problems (e.g. validation failures).
              items:
                $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required: [message]
      additionalProperties: false
      properties:
        field:
          type: string
          description: Dotted path to the offending field, when applicable.
        message:
          type: string
          description: What is wrong with this field.
  responses:
    Error:
      description: Error response.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RuntimeUnauthorized:
      description: >-
        Uniform machine-channel authentication refusal. The response must not reveal whether the bearer,
        organization, device, agent kind, or revocation state exists.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"